Add functionality in commands

+++ MyProject/Hobby_Project/Application/Categories/Commands/Delete/DeleteCategoryCommandHandler.cs
@@ -11,17 +11,27 @@ namespace Application.Categories.Commands.Delete
     public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand, int>
     {
         private readonly ICategoryRepository _categoryRepository;
+        private ILog _log;
 
         public DeleteCategoryCommandHandler(ICategoryRepository categoryRepository)
         {
             _categoryRepository = categoryRepository;
+            _log = SingletonLogger.Instance;
         }
 
         public Task<int> Handle(DeleteCategoryCommand command, CancellationToken cancellationToken)
         {
-            _categoryRepository.DeleteCategoryByID(command.Id);
-            SingletonLogger.Instance.LogMessage("delete", "Category with Id: " + command.Id + " is deleted");
-            return Task.FromResult(command.Id);
+            try
+            {
+               if (command == null) throw new NullReferenceException("Delete category command is null");
+                        
+               _categoryRepository.DeleteCategoryByID(command.Id);
+               return Task.FromResult(command.Id);
+            }catch(Exception e)
+            {
+                _log.LogError(e.Message);
+                return Task.FromResult(0);
+;           }
         }
     }
 }
